Beginnings of Vzaar-hosted HD & SD with video.js.
The first step is to host our own version of video.js, so we can use the
un-minified version [1] rather than the CDN version. This is because we
want to try using the resolutions video.js plugin [2] which doesn't work
with the obfuscated function names in the CDN version.

This commit does not include the plugin or any reference to the HD version
of the video - I plan to add that in a separate commit once I've verified
that this basic version is working.

[1] video.dev.js
[2] https://github.com/vidcaster/video-js-resolutions